fix: size_t path_sz in pick_open/save_path; add TITLEBAR_HEIGHT to filepicker dialog height
Agent-Logs-Url: https://github.com/corepunch/orion-ui/sessions/33a65e70-526a-4dc8-a51b-49b4f2864d15

+++ examples/taskmanager/view_menubar.c
@@ -55,7 +55,7 @@ static const accel_t kAccelEntries[] = {
 // File path helpers (open/save dialogs, .tdb extension)
 // ============================================================
 
-static bool pick_open_path(window_t *parent, char *path, int path_sz) {
+static bool pick_open_path(window_t *parent, char *path, size_t path_sz) {
   openfilename_t ofn = {0};
   ofn.lStructSize  = sizeof(ofn);
   ofn.hwndOwner    = parent;
@@ -67,7 +67,7 @@ static bool pick_open_path(window_t *parent, char *path, int path_sz) {
   return get_open_filename(&ofn);
 }
 
-static bool pick_save_path(window_t *parent, char *path, int path_sz) {
+static bool pick_save_path(window_t *parent, char *path, size_t path_sz) {
   openfilename_t ofn = {0};
   ofn.lStructSize  = sizeof(ofn);
   ofn.hwndOwner    = parent;
